FlaskApi/requirements.txt
Flask
Flask-SQLAlchemy
psycopg[binary]
python-dotenv==1.0.0